+++ ChangeLog.md
@@ -5,6 +5,7 @@
 
 ## Features
 * new "read" option "-n" for retrieving name/value pairs in numeric form
+* new "find" option "-e" for exactly matching name and optional circuit (ignoring case)
 * added numeric and verbose options to JSON and use "null" for unset or replacement value
 * allow passing raw value when writing name/value pairs